+++ src/utils.js
@@ -5,10 +5,8 @@ import { gzip } from 'zlib';
 import log from 'apify-shared/log';
 import retry from 'async-retry';
 import ApifyClientError, {
-    INVALID_PARAMETER_ERROR_TYPE_V1,
-    INVALID_PARAMETER_ERROR_TYPE_V2,
-    REQUEST_FAILED_ERROR_TYPE_V1,
-    REQUEST_FAILED_ERROR_TYPE_V2,
+    INVALID_PARAMETER_ERROR_TYPE,
+    REQUEST_FAILED_ERROR_TYPE,
     REQUEST_FAILED_ERROR_MESSAGE,
     NOT_FOUND_STATUS_CODE,
 } from './apify_error';
@@ -42,8 +40,7 @@ export const safeJsonParse = (str) => {
  *
  * then uses its error type or message or both.
  */
-export const newApifyClientErrorFromResponse = (body, isApiV1, details) => {
-    const REQUEST_FAILED_ERROR_TYPE = isApiV1 ? REQUEST_FAILED_ERROR_TYPE_V1 : REQUEST_FAILED_ERROR_TYPE_V2;
+export const newApifyClientErrorFromResponse = (body, details) => {
     let parsedBody = {};
 
     if (_.isObject(body)) parsedBody = body;
@@ -60,8 +57,8 @@ export const newApifyClientErrorFromResponse = (body, isApiV1, details) => {
  * Checks that given parameter is of given type and throws ApifyClientError.
  * If errorMessage is not provided then error message is created from name and type of param.
  *
- * @param value - user entered value of that parameter
- * @param {String} name - parameter name (crawlerId for options.crawlerId)
+ * @param {String} value - user entered value of that parameter
+ * @param {String} name - parameter name (actId for options.actId)
  * @param {String} type - "String", "Number", ... (see ee: https://github.com/gkz/type-check)
  * @param {String} [errorMessage] - optional error message
  * @param {Boolean} [isApiV1] - flag for legacy Crawler
@@ -69,8 +66,6 @@ export const newApifyClientErrorFromResponse = (body, isApiV1, details) => {
 export const checkParamOrThrow = (value, name, type, errorMessage, isApiV1) => {
     // TODO: move this into apify-shared along with an ApifyClientError,
     // actually it shouldn't be ApifyClientError but ApifyError in most cases!
-
-    const INVALID_PARAMETER_ERROR_TYPE = isApiV1 ? INVALID_PARAMETER_ERROR_TYPE_V1 : INVALID_PARAMETER_ERROR_TYPE_V2;
 
     if (!errorMessage) errorMessage = `Parameter "${name}" of type ${type} must be provided`;
 
@@ -136,6 +131,8 @@ export const parseBody = (body, contentType) => {
  * Wrap results from response and parse attributes from apifier headers.
  */
 export function wrapArray(response) {
+    const limit = response.headers['x-apifier-pagination-limit'] || response.headers['x-apify-pagination-limit'];
+
     /**
      * @typedef {Object} PaginationList
      * @property {Array} items - List of returned objects
@@ -149,7 +146,7 @@ export function wrapArray(response) {
         total: parseInt(response.headers['x-apifier-pagination-total'] || response.headers['x-apify-pagination-total'], 10),
         offset: parseInt(response.headers['x-apifier-pagination-offset'] || response.headers['x-apify-pagination-offset'], 10),
         count: parseInt(response.headers['x-apifier-pagination-count'] || response.headers['x-apify-pagination-count'], 10),
-        limit: parseInt(response.headers['x-apifier-pagination-limit'] || response.headers['x-apify-pagination-limit'], 10),
+        limit: limit ? parseInt(limit, 10) : null, // Limit can be null in a case of dataset items.
     };
 }