+++ test/request_queues.js
@@ -185,6 +185,63 @@ describe('RequestQueues methods', () => {
             validateRequest({}, { queueId });
         });
 
+        it('getQueue() returns null on 404 status code (RECORD_NOT_FOUND)', () => {
+            const queueId = 'some-id';
+
+            requestExpectErrorCall({
+                json: true,
+                method: 'GET',
+                url: `${BASE_URL}${BASE_PATH}/${queueId}`,
+                qs: {},
+            }, false, 404);
+
+            const apifyClient = new ApifyClient(OPTIONS);
+
+            return apifyClient
+                .requestQueues
+                .getQueue({ queueId })
+                .then(given => expect(given).to.be.eql(null));
+        });
+
+        it('updateQueue() works', () => {
+            const queueId = 'some-id';
+            const token = 'my-token';
+            const queue = { id: queueId, name: 'my-name' };
+
+            requestExpectCall({
+                json: true,
+                method: 'PUT',
+                url: `${BASE_URL}${BASE_PATH}/${queueId}`,
+                qs: { token },
+                body: _.omit(queue, 'id'),
+            });
+
+            const apifyClient = new ApifyClient(OPTIONS);
+
+            return apifyClient
+                .requestQueues
+                .updateQueue({ queueId, queue, token });
+        });
+
+        it('deleteQueue() works', () => {
+            const queueId = 'some-id';
+            const token = 'my-token';
+
+            requestExpectCall({
+                json: true,
+                method: 'DELETE',
+                url: `${BASE_URL}${BASE_PATH}/${queueId}`,
+                qs: { token },
+            });
+
+            const apifyClient = new ApifyClient(OPTIONS);
+
+            return apifyClient
+                .requestQueues
+                .deleteQueue({ queueId, token });
+        });
+
+        it('addRequest() works without forefront param', () => {
         it('addRequest() works without forefront param', async () => {
             const queueId = 'some-id';
             const request = { url: 'http://example.com' };